Minor improvements and tweaks
