+++ token/js/src/extensions/transferHook/instructions.ts
@@ -1,14 +1,15 @@
 import { struct, u8 } from '@solana/buffer-layout';
 import type { Commitment, Connection, PublicKey, Signer } from '@solana/web3.js';
 import { TransactionInstruction } from '@solana/web3.js';
-import { programSupportsExtensions, TOKEN_2022_PROGRAM_ID } from '../../constants.js';
+import { programSupportsExtensions, TOKEN_2022_PROGRAM_ID, TOKEN_PROGRAM_ID } from '../../constants.js';
 import { TokenUnsupportedInstructionError } from '../../errors.js';
 import { addSigners } from '../../instructions/internal.js';
 import { TokenInstruction } from '../../instructions/types.js';
 import { publicKey } from '@solana/buffer-layout-utils';
-import { createTransferCheckedInstruction } from '../../instructions/index.js';
-import { addExtraAccountsToInstruction } from './actions.js';
+import { createTransferCheckedInstruction } from '../../instructions/transferChecked.js';
 import { createTransferCheckedWithFeeInstruction } from '../transferFee/instructions.js';
+import { getMint } from '../../state/mint.js';
+import { getExtraAccountMetaAccount, getExtraAccountMetas, getTransferHook, resolveExtraAccountMeta } from './state.js';
 
 export enum TransferHookInstruction {
     Initialize = 0,
@@ -117,6 +118,58 @@ export function createUpdateTransferHookInstruction(
 }
 
 /**
+ * Add extra accounts needed for transfer hook to an instruction
+ *
+ * @param connection      Connection to use
+ * @param instruction     The transferChecked instruction to add accounts to
+ * @param commitment      Commitment to use
+ * @param programId       SPL Token program account
+ *
+ * @return Instruction to add to a transaction
+ */
+export async function addExtraAccountsToInstruction(
+    connection: Connection,
+    instruction: TransactionInstruction,
+    mint: PublicKey,
+    commitment?: Commitment,
+    programId = TOKEN_PROGRAM_ID
+): Promise<TransactionInstruction> {
+    if (!programSupportsExtensions(programId)) {
+        throw new TokenUnsupportedInstructionError();
+    }
+
+    const mintInfo = await getMint(connection, mint, commitment, programId);
+    const transferHook = getTransferHook(mintInfo);
+    if (transferHook == null) {
+        return instruction;
+    }
+
+    const extraAccountsAccount = getExtraAccountMetaAccount(transferHook.programId, mint);
+    const extraAccountsInfo = await connection.getAccountInfo(extraAccountsAccount, commitment);
+    if (extraAccountsInfo == null) {
+        return instruction;
+    }
+
+    const extraAccountMetas = getExtraAccountMetas(extraAccountsInfo);
+
+    const accountMetas = instruction.keys;
+    accountMetas.push({ pubkey: extraAccountsAccount, isSigner: false, isWritable: false });
+
+    for (const extraAccountMeta of extraAccountMetas) {
+        const accountMeta = resolveExtraAccountMeta(
+            extraAccountMeta,
+            accountMetas,
+            instruction.data,
+            transferHook.programId
+        );
+        accountMetas.push(accountMeta);
+    }
+    accountMetas.push({ pubkey: transferHook.programId, isSigner: false, isWritable: false });
+
+    return new TransactionInstruction({ keys: accountMetas, programId, data: instruction.data });
+}
+
+/**
  * Construct an transferChecked instruction with extra accounts for transfer hook
  *
  * @param connection            Connection to use
@@ -142,7 +195,7 @@ export async function createTransferCheckedWithTransferHookInstruction(
     decimals: number,
     multiSigners: (Signer | PublicKey)[] = [],
     commitment?: Commitment,
-    programId = TOKEN_2022_PROGRAM_ID
+    programId = TOKEN_PROGRAM_ID
 ) {
     const rawInstruction = createTransferCheckedInstruction(
         source,
@@ -194,7 +247,7 @@ export async function createTransferCheckedWithFeeAndTransferHookInstruction(
     fee: bigint,
     multiSigners: (Signer | PublicKey)[] = [],
     commitment?: Commitment,
-    programId = TOKEN_2022_PROGRAM_ID
+    programId = TOKEN_PROGRAM_ID
 ) {
     const rawInstruction = createTransferCheckedWithFeeInstruction(
         source,
